+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Schema/NamedObject.cs
@@ -1,11 +1,7 @@
 // SPDX-FileCopyrightText: 2023 Unity Technologies and the glTFast authors
 // SPDX-License-Identifier: Apache-2.0
 
-#if UNITY_6000_5_OR_NEWER
-using System.Text.Json.Serialization;
-#else
 using Unity.Gltfast.Text.Json.Serialization;
-#endif
 
 namespace GLTFast.Schema
 {